#15857: Add exp forge sweep test
